requirements.txt
backoff==2.2.1
openai==1.7.2
python-dotenv
kagglehub
transformers
torch
accelerate>=0.26.0